Update CodeIgniter Remove 'autoinit' DB setting
It doesn't make sense to do a load->database() call
but not connect to the database. IIRC there was more
stuff in CI_DB_driver::initialize() at some point,
so that was probably the reason why the setting
existed in the first place. However, now it only
results in users making invalid bug reports because
they don't understand the feature ...

Examples during just the past 2 weeks: #3571 #3601 #3607